# backend/public/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: Terminal302 API
  version: 0.1.0
  description: API REST inicial para autenticacion de Terminal302.
servers:
  - url: http://localhost:8302/api
    description: Docker local
  - url: /api
    description: Mismo host
...
paths:
...
  /forgot-password:
    post:
      tags:
        - Autenticacion
      summary: Solicitar recuperacion de contrasena
      description: Envia un enlace temporal si el correo esta registrado. La respuesta no revela si la cuenta existe.
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Solicitud procesada con respuesta generica.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: El correo no tiene un formato valido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /reset-password:
    post:
      tags:
        - Autenticacion
      summary: Restablecer contrasena
      description: Cambia la contrasena mediante un token temporal de un solo uso y revoca las sesiones existentes.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Contrasena restablecida correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Datos invalidos o enlace vencido, invalido o utilizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
  /change-initial-password:
    post:
      tags:
        - Autenticacion
      summary: Cambiar contrasena inicial obligatoria
      description: Requiere la contrasena actual y una nueva contrasena confirmada. Marca must_change_password como false.
      operationId: changeInitialPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeInitialPasswordRequest'
            examples:
              cambioInicial:
                value:
                  current_password: TEMPORAL
                  password: NuevaClave123
                  password_confirmation: NuevaClave123
      responses:
        '200':
          description: Contrasena actualizada correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeInitialPasswordResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '422':
          description: Error de validacion o contrasena actual invalida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  responses:
    Unauthenticated:
      description: Token ausente, invalido o expirado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Unauthenticated.
...
  schemas:
...
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: usuario@example.com
    ResetPasswordRequest:
      type: object
      required:
        - email
        - token
        - password
        - password_confirmation
      properties:
        email:
          type: string
          format: email
          example: usuario@example.com
        token:
          type: string
          description: Token recibido por correo, valido durante 60 minutos y de un solo uso.
        password:
          type: string
          format: password
          minLength: 8
          description: Debe contener mayuscula, minuscula, numero y simbolo.
          example: NuevaClave1!
        password_confirmation:
          type: string
          format: password
          example: NuevaClave1!
    ChangeInitialPasswordRequest:
      type: object
      required:
        - current_password
        - password
        - password_confirmation
      properties:
        current_password:
          type: string
          format: password
          example: TEMPORAL
        password:
          type: string
          format: password
          minLength: 8
          description: Debe contener letras y numeros.
          example: NuevaClave123
        password_confirmation:
          type: string
          format: password
          example: NuevaClave123
...
    ChangeInitialPasswordResponse:
      type: object
      required:
        - message
        - user
      properties:
        message:
          type: string
          example: Contrasena actualizada correctamente.
        user:
          $ref: '#/components/schemas/User'
...
    User:
      type: object
      required:
        - id
        - name
        - email
        - role
        - estado
        - must_change_password
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Administrador Terminal302
        email:
          type: string
          format: email
          example: admin@terminal302.local
        role:
          $ref: '#/components/schemas/CatalogItem'
        estado:
          $ref: '#/components/schemas/CatalogItem'
        must_change_password:
          type: boolean
          example: true
        created_at:
          type:
            - string
            - 'null'
          format: date-time
          example: '2026-06-23T19:00:00.000000Z'
        updated_at:
          type:
            - string
            - 'null'
          format: date-time
          example: '2026-06-23T19:00:00.000000Z'
    CatalogItem:
      type: object
      properties:
        id:
          type:
            - integer
            - 'null'
          example: 1
        nombre:
          type:
            - string
            - 'null'
          example: administrador
...
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Sesion cerrada correctamente.
    ValidationErrorResponse:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
          example: Las credenciales no son validas.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - Las credenciales no son validas.